Switch weight_delta calculation to int128_t instead of double since double does not guarantee a correct result for large values

+++ contracts/stakevote/stakevote.cpp
@@ -23,8 +23,8 @@ void stakevote::stakeobsv(const vector<account_stake_delta> &stake_deltas, const
     auto                         weights = weight_table{get_self(), dac_id.value};
 
     for (auto asd : stake_deltas) {
-        const auto weight_delta_s = S{asd.stake_delta.amount}.to<double>() * S{asd.unstake_delay}.to<double>() *
-                                    (S{config.time_multiplier}.to<double>() / S{time_divisor});
+        const auto weight_delta_s = S{asd.stake_delta.amount}.to<int128_t>() * S{asd.unstake_delay}.to<int128_t>() *
+                                    (S{config.time_multiplier}.to<int128_t>() / S{time_divisor});
         const int64_t weight_delta = weight_delta_s.to<int64_t>();
         auto          vw_itr       = weights.find(asd.account.value);
         if (vw_itr != weights.end()) {
@@ -98,8 +98,8 @@ void stakevote::collectwts(uint16_t batch_size, uint32_t unstake_time, name dac_
     while (stake != stakes.end() && counter < batch_size) {
         auto vw_itr = weights.find((stake->account).value);
         if (vw_itr == weights.end()) {
-            const auto weight_delta_s = S{(stake->stake).amount}.to<double>() * S{unstake_time}.to<double>() *
-                                        (S{config.time_multiplier}.to<double>() / S{time_divisor});
+            const auto weight_delta_s = S{(stake->stake).amount}.to<int128_t>() * S{unstake_time}.to<int128_t>() *
+                                        (S{config.time_multiplier}.to<int128_t>() / S{time_divisor});
             const int64_t weight_delta = weight_delta_s.to<int64_t>();
             weights.emplace(get_self(), [&](auto &v) {
                 v.voter  = stake->account;
